Update 3,4
Tratar de usar try and trow con if

+++ exercici4/Circle.h
@@ -8,9 +8,9 @@
 
 
 class Circle {
-    int area;
+    float area;
 public:
-    Circle(int a);
+    Circle(float a);
     double getArea();
 
 };
